move sms endpoints from `/profile`to `/account`
