+++ packages/manager/src/queries/profile.ts
@@ -2,11 +2,7 @@ import {
   getProfile,
   listGrants,
   Profile,
-  sendCodeToPhoneNumber,
-  SendPhoneVerificationCodePayload,
   updateProfile,
-  verifyPhoneNumberCode,
-  VerifyVerificationCodePayload,
 } from '@linode/api-v4/lib/profile';
 import { APIError } from '@linode/api-v4/lib/types';
 import { useMutation, useQuery } from 'react-query';
@@ -47,13 +43,3 @@ export const useGrants = () =>
 export const getProfileData = () => queryClient.getQueryData<Profile>(queryKey);
 export const getGrantData = () =>
   queryClient.getQueryData<Grants>(`${queryKey}-grants`);
-
-export const useSendPhoneVerificationCodeMutation = () =>
-  useMutation<{}, APIError[], SendPhoneVerificationCodePayload>(
-    sendCodeToPhoneNumber
-  );
-
-export const useVerifyPhoneVerificationCodeMutation = () =>
-  useMutation<{}, APIError[], VerifyVerificationCodePayload>(
-    verifyPhoneNumberCode
-  );
